pyproject.toml
[project]
name = "job-hunt-2"
version = "0.2.0"
description = "Automated job search with AI-powered scoring"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.32.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "questionary>=2.0.0",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "PyPDF2>=3.0.0",
]
